refactor: 폼 분리 제안 사항
