+++ src/pages/GatheringCreate/components/GatheringCreateForm/useCreateGathering.ts
@@ -1,0 +1,23 @@
+import {
+  GatheringCreateRequest,
+  usePostGatheringCreateApi,
+} from '@/apis/postGatheringCreate';
+import { showErrorToast } from '@/utils/showToast';
+
+import { OnGatheringCreate } from '.';
+
+const useCreateGathering = (onCreate: OnGatheringCreate) => {
+  const gatheringCreateApi = usePostGatheringCreateApi();
+
+  return async (request: GatheringCreateRequest) => {
+    const { data, isBadRequest } = await gatheringCreateApi(request);
+
+    if (isBadRequest) {
+      return showErrorToast(data.message);
+    }
+
+    onCreate(data.roomId);
+  };
+};
+
+export default useCreateGathering;
